Make sure the latest commands image is used
We need to try to pull the image first and then fall back to getting
the local image. The image will either be available remotely and will
be `latest` or an immutable tag, in which case we want to make sure
we have it, or will be local only and the pull will fail so we fall
back to getting it.