Remove additional Error or Done from cdflow output
This removes the string "Error" from the output, which was being
displayed with the usage message, and removes the "Done" string
otherwise - the user being presented with a prompt and a zero exit
status should be enough to signal that the process is done. This also
makes it a thinner wrapper around the cdflow container.